Fix fixture naming
